Split up models in sensor.core into multiple files
