+++ sensor/core/models/__init__.py
@@ -1,0 +1,6 @@
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+from sensor import SensorType, GenericSensor, Sensor
+from event import GenericEvent, Event
