+++ sensor/core/models/sensor.py
@@ -32,22 +32,3 @@ class Sensor(models.Model):
 
     class Meta:
         abstract = True
-
-
-class GenericEvent(models.Model):
-    """Represents a measurement event abstracting away what exactly is
-    measured.
-    """
-
-    sensor = models.ForeignKey('core.GenericSensor')
-    datetime = models.DateTimeField()
-    value = models.CharField(max_length=128)
-
-
-class Event(models.Model):
-    """Base class for sensor-specific event types"""
-
-    generic_event = models.OneToOneField('core.GenericEvent')
-
-    class Meta:
-        abstract = True
